clean up pipelines module now that we support so many different kind of pipelines
