+++ src/pipelines/fill_mask/mod.rs
@@ -1,0 +1,5 @@
+// Fill mask pipeline module
+pub mod pipeline;
+
+// Re-export for convenience
+pub use pipeline::*;
