+++ src/pipelines/fill_mask/pipeline.rs
