+++ src/pipelines/sentiment_analysis/mod.rs
@@ -1,0 +1,5 @@
+// Sentiment analysis pipeline module
+pub mod pipeline;
+
+// Re-export for convenience
+pub use pipeline::*;
