+++ src/pipelines/text_generation/capabilities.rs
