+++ src/pipelines/text_generation/combined_pipelines.rs
@@ -1,21 +1,26 @@
 use super::{
     tools::{Tool, ToolCallResult},
-    traits::{TextGenerationPipeline, TextGenerationPipelineWithTools},
-    TextGenerationModel,
+    traits::{
+        TextGenerationPipeline, TextGenerationPipelineWithToggleableReasoning,
+        TextGenerationPipelineWithTools,
+    },
 };
+use crate::pipelines::TextGenerationModel;
 use crate::Message;
 use tokenizers::Tokenizer;
 
-/// A text generation pipeline with tool calling capabilities.
-/// Used for models like Gemma3 that can register and use tools during generation.
-pub struct ToolCallingPipeline {
+/// A text generation pipeline with both toggleable reasoning and tool calling capabilities.
+/// Used for models like Qwen3 that support both reasoning control and tool use.
+pub struct ToggleableReasoningToolsPipeline {
     model: Box<dyn TextGenerationModel>,
     tokenizer: Tokenizer,
     eos_token_id: u32,
+    reasoning_enabled: bool,
     tools: Vec<Tool>,
+    reasoning_trace: Option<String>,
 }
 
-impl ToolCallingPipeline {
+impl ToggleableReasoningToolsPipeline {
     pub fn new(
         model: Box<dyn TextGenerationModel>,
         tokenizer: Tokenizer,
@@ -25,15 +30,21 @@ impl ToolCallingPipeline {
             model,
             tokenizer,
             eos_token_id,
+            reasoning_enabled: false, // Default to disabled
             tools: Vec::new(),
+            reasoning_trace: None,
         }
     }
 }
 
-impl TextGenerationPipeline for ToolCallingPipeline {
+impl TextGenerationPipeline for ToggleableReasoningToolsPipeline {
     fn prompt_completion(&self, prompt: &str, max_length: usize) -> anyhow::Result<String> {
-        // Format the prompt
-        let formatted_prompt = self.model.format_prompt(prompt);
+        // Use reasoning-aware prompt formatting if reasoning is enabled
+        let formatted_prompt = if self.reasoning_enabled {
+            self.model.format_prompt_with_reasoning(prompt, true)
+        } else {
+            self.model.format_prompt(prompt)
+        };
 
         // Turn the prompt into tokens
         let prompt_tokens = self
@@ -88,7 +99,25 @@ impl TextGenerationPipeline for ToolCallingPipeline {
     }
 }
 
-impl TextGenerationPipelineWithTools for ToolCallingPipeline {
+impl TextGenerationPipelineWithToggleableReasoning for ToggleableReasoningToolsPipeline {
+    fn enable_reasoning(&mut self) {
+        todo!("Reasoning enabling not yet implemented")
+    }
+
+    fn disable_reasoning(&mut self) {
+        todo!("Reasoning disabling not yet implemented")
+    }
+
+    fn is_reasoning_enabled(&self) -> bool {
+        todo!("Reasoning status check not yet implemented")
+    }
+
+    fn get_reasoning_trace(&self) -> Option<&str> {
+        todo!("Reasoning trace access not yet implemented")
+    }
+}
+
+impl TextGenerationPipelineWithTools for ToggleableReasoningToolsPipeline {
     fn register_tool(&mut self, tool: Tool) -> anyhow::Result<()> {
         todo!("Tool registration not yet implemented")
     }
