+++ src/pipelines/text_generation/tools.rs
