+++ src/pipelines/text_generation/traits.rs
