+++ src/pipelines/zero_shot_classification/mod.rs
@@ -1,0 +1,5 @@
+// Zero-shot classification pipeline module
+pub mod pipeline;
+
+// Re-export for convenience
+pub use pipeline::*;
