+++ src/pipelines/zero_shot_classification/pipeline.rs
